requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx==0.25.2
python-dotenv==1.0.0
SQLAlchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.13.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
